fix: cleaning and refactoring
